카메라가 방에서 벗어나는거 잠시 폐기

+++ ISAAC/APIClass/CPanelUI.cpp
@@ -36,7 +36,7 @@ void CPanelUI::render(HDC _dc)
 
 	Vec2 vPos = GetPos();
 
-	TransparentBlt(_dc, vPos.x, vPos.y
+	TransparentBlt(_dc, (float)vPos.x, (float)vPos.y
 		, GetIdleTex()->Widht()
 		, GetIdleTex()->Height()
 		, GetIdleTex()->GetDC()
